requirements.txt
python-dotenv==1.0.1
crewai==0.26.11
anthropic==0.25.7
